Extract `scalac` options checks from `SipScalaTests` to the `compile` sub-command test suite
